bagfix
перенес данные из папки views в папку pages
